+++ lib/utils/hero_tag.dart
@@ -1,3 +1,0 @@
-class EditHeroTag {
-  const EditHeroTag();
-}